requirements.txt
bitstring==4.0.2
cffi==1.15.1
cryptography==41.0.3
ecdsa==0.18.0
esptool==4.6.2
ftd2xx==1.3.3
pycparser==2.21
pyserial==3.5
pyusb==1.2.1
PyYAML==6.0.1
reedsolo==1.7.0
six==1.16.0
